Normalized and split logic, corrected and added docs, fixed standards (#6)
